feat: Export a summarizer interface to allow custom summarizer implementations
