+++ packages/istanbul-lib-report/summarizer.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./lib/summarizer');
